Complete the html functionality for making a turn
---Major Points---
1) This setup can now handle every aspect of making a move:
    - The chosen orbs are swapped
    - The matches are deleted
    - The unaffected orbs drop down
    - Attic orbs drop down to fill in the space
2) `animationHelpers.js` has been added
    - This helps sort out the functions that the demo uses. Anything
    that affects the html is now inside `animationHelpers.js`
3) There is a new logic (again) for pulling down unaffected orbs

---Details---
- In general, each function only does what it says. For example,
`swap()`, which is now inside `animationHelpers.js` used to call
`repopulate()` after each swap due to the previous pull down logic,
but now it only swaps.

- `pullDownOrbs()` has been split into three functions
    - `eraseMatches(matches)`
        - gives each orb inside the matches an opacity of 0
    - `activateGravity()`
        - pulls down the unaffected orbs with NEW LOGIC:
            1) Iterates through each orb, again starting from the
            second to last row going left to right, bottom to top.
            2) For each orb, it checks the opacity of each orb below,
            and tallies the blank count.
            3) If an orb has one or more blanks below it, then it is
            added to a blanksBelow object that has the orb
            coordinates as a key and the blank count as a value.
            For example:
            ```
            blanksBelow === {
                [0, 1]: 1,
                [0, 2]: 1,
                [0, 3]: 1
            }
            ```
            4) For each set of coords(key) in blanksBelow, that orb
            is moved down blank count(value) spaces.
            *This logic is way better and simpler and should
            translate to `orbs.js` very easily
    - `releaseAttic(orbCounts)`
        - drops attic orbs in to fill in the empty space.
        - The logic for this is the same based on an orbCounts
        object. The only difference is that it is calculated based
        on `board.matches` instead of being returned by
        `board.evaluate` as was originally intended.

- `makeMove(row, col, direction, distance)`
    - takes an orb and moves it where it needs to go!
    - used by other functions like `swap`
