+++ site/scripts/main.js
@@ -2,6 +2,7 @@
 
 let board = require('../../src/board');
 let demotools = require('./demotools');
+let animate = require('./animationHelpers');
 let _ = require('lodash');
 
 let colors = ['red', 'blue', 'green', 'aquamarine', 'yellow', 'brown', 'purple', 'orange'];
@@ -9,12 +10,12 @@ let b = new board.Board(8, 8, colors);
 
 // create the main board and append it to 'board' div
 let demoBoard = document.getElementById('board');
-demotools.createHTMLBoard(b, demoBoard, 'main');
+animate.createHTMLBoard(b, demoBoard, 'main');
 
 // create the attic board and append it to 'attic' div
 let a = new board.Board(8, 8, colors);
 let atticBoard = document.getElementById('attic');
-demotools.createHTMLBoard(a, atticBoard, 'attic');
+animate.createHTMLBoard(a, atticBoard, 'attic');
 
 // create a scoreboard for each orb type
 let types = document.getElementById('types');
@@ -39,54 +40,24 @@ global.selectOrb = function(row, col) {
     // if both selected orbs are the same, deselect
     // if the two orbs aren't neighbors, throw error and reset choices
     if (selectedOrbs.length === 1) {
-        demotools.addBorder(row, col);
+        animate.addBorder(row, col);
         return;
     } else if (demotools.areNeighbors(selectedOrbs)) {
         let matchDatas = demotools.makeMove(b, selectedOrbs);
         _.each(matchDatas, matchData => {
             _.each(matchData, match => {
-                demotools.updateScore(match);
+                animate.updateScore(match);
             })
         });
         selectedOrbs = [];
     } else if (demotools.twoSelectedOrbsAreEqual(selectedOrbs)) {
         selectedOrbs = [];
-        demotools.removeBorder(row, col);
+        animate.removeBorder(row, col);
     } else {
         alert("You must choose two neighboring orbs!");
         _.each(selectedOrbs, orb => {
-            demotools.removeBorder(orb[0], orb[1]);
+            animate.removeBorder(orb[0], orb[1]);
         })
         selectedOrbs = [];
     }
-}
-
-// practice the movements
-global.move = function() {
-    let matchOrb = document.getElementById('main 7 0');
-    matchOrb.style.opacity = '0';
-    _.each(_.range(8), row => {
-        let orb = document.getElementById('main ' + row + ' 0');
-        orb.style.webkitTransform = 'translate(0, 56px)';
-    })
-    let atticOrb = document.getElementById('attic 7 0');
-    atticOrb.style.webkitTransform = 'translate(0, 112px)';
-}
-
-// working backwards, here is the example output from the new-to-be Board.evaluate()
-let exampleOrbCounts = {
-    1: 1,
-    2: 1,
-    3: 1
-};
-let exampleMatches = [
-    [[1, 1], [1, 2], [1, 3]]
-];
-
-global.swap = function() {
-    demotools.swap(b, [4, 5], [5, 5]);
-}
-
-global.moveMatch = function() {
-    demotools.pullDownOrbs(b, exampleMatches, exampleOrbCounts);
 }
